Убрать неиспользуемый FilmRowMapper и переименовать FilmRowMapperAllFields в FilmRowMapper
